per-file: Add a search option
This for now searches for just files/folders by name, the exact search
behavior can be refined later when it is clear what is useful.

To prevent the browser from locking up there are two protections,
first we only search if at least three codepoints are entered secondly
we only display the first 500 results.